Moved subsystem related apis in subsystem blueprint from app.py
